Add a `contentEncoding` property to compressed files.
This can be used later on by other parts of a pipeline to:
 * Tell that data was indeed compressed (without brittle extension checking),
 * Tell how the data was compressed.

It follows roughly the same semantics as the HTTP 'Content-Encoding' header and allows a consumer of the file to observe all the encoding processes that occurred on a file. See: http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html section 14.11.